+++ packages/thirdweb/src/adapters/ethers5.test.ts
@@ -1,8 +1,9 @@
 import { describe, test, expect } from "vitest";
 import { privateKeyAccount } from "../wallets/private-key.js";
 import { TEST_CLIENT } from "../../test/src/test-clients.js";
-import { ethers6Adapter } from "./ethers6.js";
+import { toEthersSigner } from "./ethers5.js";
 import { ANVIL_CHAIN } from "../../test/src/chains.js";
+import * as ethers5 from "ethers5";
 import { ANVIL_PKEY_A, TEST_ACCOUNT_B } from "../../test/src/test-wallets.js";
 
 const account = privateKeyAccount({
@@ -10,35 +11,38 @@ const account = privateKeyAccount({
   client: TEST_CLIENT,
 });
 
-describe("toEthersSigner", () => {
-  test("should return an ethers 6 signer", async () => {
-    const signer = await ethers6Adapter.signer.toEthers({
-      client: TEST_CLIENT,
+describe("ethers5 adapter", () => {
+  test("should return an ethers 5 signer", async () => {
+    const signer = await toEthersSigner(
+      ethers5,
+      TEST_CLIENT,
       account,
-      chain: ANVIL_CHAIN,
-    });
+      ANVIL_CHAIN,
+    );
     expect(signer).toBeDefined();
     expect(signer.signMessage).toBeDefined();
   });
 
   test("should sign message", async () => {
-    const signer = await ethers6Adapter.signer.toEthers({
-      client: TEST_CLIENT,
+    const signer = await toEthersSigner(
+      ethers5,
+      TEST_CLIENT,
       account,
-      chain: ANVIL_CHAIN,
-    });
+      ANVIL_CHAIN,
+    );
     const expectedSig = await account.signMessage({ message: "hello world" });
     const sig = await signer.signMessage("hello world");
     expect(sig).toBe(expectedSig);
   });
 
   test("should sign typed data", async () => {
-    const signer = await ethers6Adapter.signer.toEthers({
-      client: TEST_CLIENT,
+    const signer = await toEthersSigner(
+      ethers5,
+      TEST_CLIENT,
       account,
-      chain: ANVIL_CHAIN,
-    });
-    expect(signer.signTypedData).toBeDefined();
+      ANVIL_CHAIN,
+    );
+    expect(signer._signTypedData).toBeDefined();
 
     // All properties on a domain are optional
     const domain = {
@@ -74,17 +78,18 @@ describe("toEthersSigner", () => {
       contents: "Hello, Bob!",
     };
 
-    const signature = await signer.signTypedData(domain, types, value);
+    const signature = await signer._signTypedData(domain, types, value);
 
     expect(signature.length).toBe(132);
   });
 
   test("should send a tx", async () => {
-    const signer = await ethers6Adapter.signer.toEthers({
-      client: TEST_CLIENT,
+    const signer = await toEthersSigner(
+      ethers5,
+      TEST_CLIENT,
       account,
-      chain: ANVIL_CHAIN,
-    });
+      ANVIL_CHAIN,
+    );
     const txResponse = await signer.sendTransaction({
       to: TEST_ACCOUNT_B.address,
       value: 100,