+++ legacy_packages/react/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @thirdweb-dev/react
 
+## 4.4.31
+
+### Patch Changes
+
+- Updated dependencies [[`b3fca95`](https://github.com/thirdweb-dev/js/commit/b3fca9509feb0ac6e0ded816f08570b2954f6521)]:
+  - @thirdweb-dev/chains@0.1.88
+  - @thirdweb-dev/react-core@4.4.31
+  - @thirdweb-dev/sdk@4.0.57
+  - @thirdweb-dev/wallets@2.4.33
+
 ## 4.4.30
 
 ### Patch Changes